ai-services/requirements.txt
fastapi
uvicorn
google-generativeai
langchain-google-genai
python-dotenv
pydantic
python-multipart
